+++ NEWS.md
@@ -25,6 +25,7 @@ functionality in `Derivate()`
 - New function `MakeBreaks()`.
 - New guide. `guide_colorstrip()` displays discretized values of a continuous color 
 or fill scale. 
+- Fix for unconnected contours in `stat_contour_fill()`
 
 # meteoR 0.0.9100
 - Full (not perfect) documentation.